Select column by user

metadata_user_select_column is used to select which columns in a dataframe/tibble corresponds to the variable of interest. It is used to compile the metadata yaml file by prompting the user to choose the relevant columns. It is used in metadata_add_locations and metadata_create_template.

metadata_user_select_column(column, choices)

Arguments

column

Name of the variable of interest

choices

The options that can be selected from